Converted GroupMenuWrapper.js to a functional component
