+++ src/components/GroupMenuWrapper.js
@@ -7,32 +7,27 @@ import BlockIcon from '@material-ui/icons/Block';
 import ModeCommentIcon from '@material-ui/icons/ModeComment';
 import EditIcon from '@material-ui/icons/Edit';
 import {
-  GroupedMenu, generateMenuData, generateMenuItem, InvalidatedIcon, CheckIcon,
+  GroupedMenu,
+  generateMenuData,
+  generateMenuItem,
+  InvalidatedIcon,
+  CheckIcon,
 } from 'tc-ui-toolkit';
+import { generateItemId } from '../helpers/groupMenuHelpers'
 
-export function generateItemId(occurrence, bookId, chapter, verse, quote) {
-  let quoteId = '';
-
-  if (Array.isArray(quote)) { // is a bit more complicated
-    const parts = [];
-
-    for (let i = 0, l = quote.length; i < l; i++) {
-      const quotePart = quote[i];
-      parts.push(quotePart.occurrence + ':' + quotePart.word);
-    }
-    quoteId = parts.join(':');
-  } else {
-    quoteId = `${occurrence}:${quote}`;
-  }
-  return `${quoteId}:${verse}:${chapter}:${bookId}`;
-}
-
-class GroupMenuWrapper extends React.Component {
+function GroupMenuWrapper({
+  translate,
+  tc: {
+    contextId,
+    groupsDataReducer: { groupsData },
+    groupsIndexReducer: { groupsIndex },
+  },
+}) {
   /**
    * Handles click events from the menu
    * @param {object} contextId - the menu item's context id
    */
-  handleClick = ({ contextId }) => {
+  function handleClick({ contextId }) {
     const { tc: { actions: { changeCurrentContextId } } } = this.props;
     changeCurrentContextId(contextId);
   };
@@ -42,7 +37,7 @@ class GroupMenuWrapper extends React.Component {
    * @param {object} item - an item in the groups data
    * @returns {object} the updated item
    */
-  onProcessItem = item => {
+  function onProcessItem(item) {
     const { tc: { project } } = this.props;
     const bookName = project.getBookName();
 
@@ -80,110 +75,105 @@ class GroupMenuWrapper extends React.Component {
     };
   };
 
-  render() {
-    const {
-      translate,
-      tc: {
-        contextId,
-        groupsDataReducer: { groupsData },
-        groupsIndexReducer: { groupsIndex },
-      },
-    } = this.props;
-    const filters = [
-      {
-        label: translate('menu.invalidated'),
-        key: 'invalidated',
-        icon: <InvalidatedIcon/>,
-      },
-      {
-        label: translate('menu.bookmarks'),
-        key: 'reminders',
-        icon: <BookmarkIcon/>,
-      },
-      {
-        label: translate('menu.selected'),
-        key: 'finished',
-        disables: ['not-finished'],
-        icon: <CheckIcon/>,
-      },
-      {
-        label: translate('no_selection_needed'),
-        key: 'nothingToSelect',
-        icon: <CheckIcon/>,
-      },
-      {
-        label: translate('menu.no_selection'),
-        id: 'not-finished',
-        key: 'finished',
-        value: false,
-        disables: ['finished'],
-        icon: <BlockIcon/>,
-      },
-      {
-        label: translate('menu.verse_edit'),
-        key: 'verseEdits',
-        icon: <EditIcon/>,
-      },
-      {
-        label: translate('menu.comments'),
-        key: 'comments',
-        icon: <ModeCommentIcon/>,
-      },
-    ];
-
-    const statusIcons = [
-      {
-        key: 'invalidated',
-        icon: <InvalidatedIcon style={{ color: 'white' }}/>,
-      },
-      {
-        key: 'reminders',
-        icon: <BookmarkIcon style={{ color: 'white' }}/>,
-      },
-      {
-        key: 'finished',
-        icon: <CheckIcon style={{ color: '#58c17a' }}/>,
-      },
-      {
-        key: 'nothingToSelect',
-        icon: <CheckIcon style={{ color: '#58c17a' }}/>,
-      },
-      {
-        key: 'verseEdits',
-        icon: <EditIcon style={{ color: 'white' }}/>,
-      },
-      {
-        key: 'comments',
-        icon: <ModeCommentIcon style={{ color: 'white' }}/>,
-      },
-    ];
-
-    const entries = generateMenuData(
-      groupsIndex,
-      groupsData,
-      'selections',
-      this.onProcessItem,
-      'nothingToSelect'
-    );
-
-    const activeEntry = generateMenuItem(contextId, this.onProcessItem);
-    const sorted = entries.sort((a, b) => {
+  function sortEntries(entries) {
+    return entries.sort((a, b) => {
       const aName = (a.title || a.id).toLowerCase();
       const bName = (b.title || b.id).toLowerCase();
       return (aName < bName) ? -1 : (aName > bName) ? 1 : 0;
     });
-    return (
-      <GroupedMenu
-        filters={filters}
-        entries={sorted}
-        active={activeEntry}
-        statusIcons={statusIcons}
-        emptyNotice={translate('menu.no_results')}
-        title={translate('menu.menu')}
-        onItemClick={this.handleClick}
-      />
-    );
   }
+
+  const filters = [
+    {
+      label: translate('menu.invalidated'),
+      key: 'invalidated',
+      icon: <InvalidatedIcon/>,
+    },
+    {
+      label: translate('menu.bookmarks'),
+      key: 'reminders',
+      icon: <BookmarkIcon/>,
+    },
+    {
+      label: translate('menu.selected'),
+      key: 'finished',
+      disables: ['not-finished'],
+      icon: <CheckIcon/>,
+    },
+    {
+      label: translate('no_selection_needed'),
+      key: 'nothingToSelect',
+      icon: <CheckIcon/>,
+    },
+    {
+      label: translate('menu.no_selection'),
+      id: 'not-finished',
+      key: 'finished',
+      value: false,
+      disables: ['finished'],
+      icon: <BlockIcon/>,
+    },
+    {
+      label: translate('menu.verse_edit'),
+      key: 'verseEdits',
+      icon: <EditIcon/>,
+    },
+    {
+      label: translate('menu.comments'),
+      key: 'comments',
+      icon: <ModeCommentIcon/>,
+    },
+  ];
+
+  const statusIcons = [
+    {
+      key: 'invalidated',
+      icon: <InvalidatedIcon style={{ color: 'white' }}/>,
+    },
+    {
+      key: 'reminders',
+      icon: <BookmarkIcon style={{ color: 'white' }}/>,
+    },
+    {
+      key: 'finished',
+      icon: <CheckIcon style={{ color: '#58c17a' }}/>,
+    },
+    {
+      key: 'nothingToSelect',
+      icon: <CheckIcon style={{ color: '#58c17a' }}/>,
+    },
+    {
+      key: 'verseEdits',
+      icon: <EditIcon style={{ color: 'white' }}/>,
+    },
+    {
+      key: 'comments',
+      icon: <ModeCommentIcon style={{ color: 'white' }}/>,
+    },
+  ];
+
+  const entries = generateMenuData(
+    groupsIndex,
+    groupsData,
+    'selections',
+    this.onProcessItem,
+    'nothingToSelect'
+  );
+
+  const activeEntry = generateMenuItem(contextId, this.onProcessItem);
+  const sorted = sortEntries(entries);
+
+  return (
+    <GroupedMenu
+      filters={filters}
+      entries={sorted}
+      active={activeEntry}
+      statusIcons={statusIcons}
+      emptyNotice={translate('menu.no_results')}
+      title={translate('menu.menu')}
+      onItemClick={this.handleClick}
+    />
+  );
 }
 
 GroupMenuWrapper.propTypes = {
